Refactor transport modules to move dummy Web3 to shared
