+++ shared/src/transport.rs
@@ -1,14 +1,13 @@
-use crate::http_transport::HttpTransport;
-use derivative::Derivative;
-use ethcontract::jsonrpc::types::{Call, Value};
-use ethcontract::web3::{error, BatchTransport, RequestId, Transport};
-use futures::future::BoxFuture;
-use futures::FutureExt;
-use std::{
-    convert::TryInto,
-    sync::Arc,
-    time::{Duration, Instant},
+pub mod dummy;
+pub mod http;
+pub mod instrumented;
+
+use self::{
+    http::HttpTransport,
+    instrumented::{MetricTransport, TransportMetrics},
 };
+use ethcontract::web3::Transport;
+use std::{convert::TryInto as _, sync::Arc, time::Duration};
 
 /// Convenience method to create our standard instrumented transport
 pub fn create_instrumented_transport<T>(
@@ -20,6 +19,11 @@ where
     <T as Transport>::Out: Send + 'static,
 {
     MetricTransport::new(transport, metrics)
+}
+
+struct NoopTransportMetrics;
+impl TransportMetrics for NoopTransportMetrics {
+    fn report_query(&self, _: &str, _: Duration) {}
 }
 
 /// Convenience method to create a compatible transport without metrics (noop)
@@ -37,75 +41,4 @@ where
     let env = std::env::var("NODE_URL").unwrap();
     let transport = HttpTransport::new(env.parse().unwrap());
     MetricTransport::new(transport, Arc::new(NoopTransportMetrics))
-}
-
-pub trait TransportMetrics: Send + Sync {
-    fn report_query(&self, label: &str, elapsed: Duration);
-}
-#[derive(Clone, Derivative)]
-#[derivative(Debug)]
-pub struct MetricTransport<T: Transport> {
-    inner: T,
-    #[derivative(Debug = "ignore")]
-    metrics: Arc<dyn TransportMetrics>,
-}
-
-impl<T: Transport> MetricTransport<T> {
-    pub fn new(inner: T, metrics: Arc<dyn TransportMetrics>) -> MetricTransport<T> {
-        Self { inner, metrics }
-    }
-}
-
-impl<T> Transport for MetricTransport<T>
-where
-    T: Transport,
-    <T as Transport>::Out: Send + 'static,
-{
-    type Out = BoxFuture<'static, error::Result<Value>>;
-
-    fn prepare(&self, method: &str, params: Vec<Value>) -> (RequestId, Call) {
-        self.inner.prepare(method, params)
-    }
-
-    fn send(&self, id: RequestId, request: Call) -> Self::Out {
-        let metrics = self.metrics.clone();
-        let start = Instant::now();
-        self.inner
-            .send(id, request.clone())
-            .inspect(move |_| {
-                let label = match request {
-                    Call::MethodCall(method) => method.method,
-                    Call::Notification(notification) => notification.method,
-                    Call::Invalid { .. } => "invalid".into(),
-                };
-                metrics.report_query(&label, start.elapsed());
-            })
-            .boxed()
-    }
-}
-
-impl<T> BatchTransport for MetricTransport<T>
-where
-    T: BatchTransport,
-    T::Batch: Send + 'static,
-    <T as Transport>::Out: Send + 'static,
-{
-    type Batch = BoxFuture<'static, error::Result<Vec<error::Result<Value>>>>;
-
-    fn send_batch<I>(&self, requests: I) -> Self::Batch
-    where
-        I: IntoIterator<Item = (RequestId, Call)>,
-    {
-        let metrics = self.metrics.clone();
-        let start = Instant::now();
-        self.inner
-            .send_batch(requests)
-            .inspect(move |_| metrics.report_query(&"batch", start.elapsed()))
-            .boxed()
-    }
-}
-
-struct NoopTransportMetrics;
-impl TransportMetrics for NoopTransportMetrics {
-    fn report_query(&self, _: &str, _: Duration) {}
 }
