+++ solver/src/interactions/uniswap.rs
@@ -57,7 +57,7 @@ impl UniswapInteraction {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::testutil;
+    use shared::transport::dummy;
     use contracts::IUniswapLikeRouter;
     use hex_literal::hex;
 
@@ -74,9 +74,9 @@ mod tests {
         let token_in = H160::from_low_u64_be(7);
         let token_out = 8;
         let payout_to = 9u8;
-        let router = IUniswapLikeRouter::at(&testutil::dummy_web3(), H160::from_low_u64_be(4));
+        let router = IUniswapLikeRouter::at(&dummy::web3(), H160::from_low_u64_be(4));
         let settlement = GPv2Settlement::at(
-            &testutil::dummy_web3(),
+            &dummy::web3(),
             H160::from_low_u64_be(payout_to as u64),
         );
         let interaction = UniswapInteraction {
