+++ solver/src/interactions/weth.rs
@@ -40,12 +40,12 @@ impl Interaction for UnwrapWethInteraction {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::testutil;
     use hex_literal::hex;
+    use shared::dummy_contract;
 
     #[test]
     fn encode_unwrap_weth() {
-        let weth = testutil::dummy_weth([0x42; 20]);
+        let weth = dummy_contract!(WETH9, [0x42; 20]);
         let amount = U256::from(13_370_000_000_000_000_000u128);
         let interaction = UnwrapWethInteraction {
             weth: weth.clone(),
@@ -66,11 +66,11 @@ mod tests {
     #[test]
     fn merge_same_native_token() {
         let mut unwrap0 = UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x01; 20]),
+            weth: dummy_contract!(WETH9, [0x01; 20]),
             amount: 1.into(),
         };
         let unwrap1 = UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x01; 20]),
+            weth: dummy_contract!(WETH9, [0x01; 20]),
             amount: 2.into(),
         };
 
@@ -81,11 +81,11 @@ mod tests {
     #[test]
     fn merge_different_native_token() {
         let mut unwrap0 = UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x01; 20]),
+            weth: dummy_contract!(WETH9, [0x01; 20]),
             amount: 1.into(),
         };
         let unwrap1 = UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x02; 20]),
+            weth: dummy_contract!(WETH9, [0x02; 20]),
             amount: 2.into(),
         };
 
@@ -97,11 +97,11 @@ mod tests {
     #[should_panic]
     fn merge_u256_overflow() {
         let mut unwrap0 = UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x01; 20]),
+            weth: dummy_contract!(WETH9, [0x01; 20]),
             amount: 1.into(),
         };
         let unwrap1 = UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x01; 20]),
+            weth: dummy_contract!(WETH9, [0x01; 20]),
             amount: U256::max_value(),
         };
