+++ solver/src/liquidity.rs
@@ -80,9 +80,10 @@ impl Settleable for LimitOrder {
 impl From<Order> for LimitOrder {
     fn from(order: Order) -> Self {
         use self::offchain_orderbook::normalize_limit_order;
-        use crate::testutil;
+        use shared::dummy_contract;
+        use contracts::WETH9;
 
-        let native_token = testutil::dummy_weth(H160([0x42; 20]));
+        let native_token = dummy_contract!(WETH9, H160([0x42; 20]));
         normalize_limit_order(order, native_token)
     }
 }
