+++ solver/src/main.rs
@@ -6,7 +6,7 @@ use reqwest::Url;
 use shared::{
     bad_token::list_based::ListBasedDetector,
     current_block::current_block_stream,
-    http_transport::HttpTransport,
+    transport::http::HttpTransport,
     maintenance::{Maintaining, ServiceMaintenance},
     metrics::serve_metrics,
     network::network_name,
