+++ solver/src/metrics.rs
@@ -8,7 +8,7 @@ use model::order::Order;
 use prometheus::{
     HistogramOpts, HistogramVec, IntCounter, IntCounterVec, IntGaugeVec, Opts, Registry,
 };
-use shared::{pool_cache::PoolCacheMetrics, transport::TransportMetrics};
+use shared::{pool_cache::PoolCacheMetrics, transport::instrumented::TransportMetrics};
 use strum::{AsStaticRef, VariantNames};
 
 use crate::liquidity::Liquidity;
