+++ solver/src/settlement/settlement_encoder.rs
@@ -282,10 +282,12 @@ impl SettlementEncoder {
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use crate::{encoding::EncodedInteraction, settlement::NoopInteraction, testutil};
+    use crate::{encoding::EncodedInteraction, settlement::NoopInteraction};
+    use contracts::WETH9;
     use ethcontract::Bytes;
     use maplit::hashmap;
     use model::order::{OrderBuilder, OrderCreation};
+    use shared::dummy_contract;
 
     #[test]
     pub fn encode_trades_finds_token_index() {
@@ -319,7 +321,7 @@ pub mod tests {
 
     #[test]
     fn settlement_merges_unwraps_for_same_token() {
-        let weth = testutil::dummy_weth([0x42; 20]);
+        let weth = dummy_contract!(WETH9, [0x42; 20]);
 
         let mut encoder = SettlementEncoder::new(HashMap::new());
         encoder.add_unwrap(UnwrapWethInteraction {
@@ -345,11 +347,11 @@ pub mod tests {
     fn settlement_encoder_appends_unwraps_for_different_tokens() {
         let mut encoder = SettlementEncoder::new(HashMap::new());
         encoder.add_unwrap(UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x01; 20]),
+            weth: dummy_contract!(WETH9, [0x01; 20]),
             amount: 1.into(),
         });
         encoder.add_unwrap(UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x02; 20]),
+            weth: dummy_contract!(WETH9, [0x02; 20]),
             amount: 2.into(),
         });
 
@@ -367,7 +369,7 @@ pub mod tests {
     fn settlement_unwraps_after_execution_plan() {
         let interaction: EncodedInteraction = (H160([0x01; 20]), 0.into(), Bytes(Vec::new()));
         let unwrap = UnwrapWethInteraction {
-            weth: testutil::dummy_weth([0x01; 20]),
+            weth: dummy_contract!(WETH9, [0x01; 20]),
             amount: 1.into(),
         };
 
@@ -444,8 +446,7 @@ pub mod tests {
 
     #[test]
     fn merge_ok() {
-        let web3 = testutil::dummy_web3();
-        let weth = contracts::WETH9::at(&web3, H160::zero());
+        let weth = dummy_contract!(WETH9, H160::zero());
 
         let prices = hashmap! { token(1) => 1.into(), token(3) => 3.into() };
         let mut encoder0 = SettlementEncoder::new(prices);
